7_LLM_and_Prompt_Engineering/pyproject.toml
[tool.poetry]
name = "7-llm-and-prompt-engineering"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
langchain-google-genai = "^2.1.3"
python-dotenv = "^1.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
